Respect the time zone of each conference
When we check the schedule of a conference, we expect the time to be
displayed in the time zone of the conference location. The conference
card and list views have been changed to properly display the talk time,
using the conference time zone.
Time zone abbreviation is also explicitly mentioned in the card views.