Use list view to show the conference schedule
When a conference hosts many talks, having a List View of Cards to show
the talks of the conference is not convenient for the users. On top of
that, the display suffers in terms of performance. Thus, the schedule
view has been substituted with a ListView, having a lightweight
AbstractListItem as delegate.