Move Event Info actions from Card to Page
Since a single Card is displayed on the Event Info page, the actions (navigate to page, add to
favorites, delete from favorites) should belong to the Page and not to the
Card. That way they are more accessible to the user -belonging to the
toolbar- and offer a more consistent UI philosophy.