Add support for adding additional indoor map content
The bulk of the work for this is in KOSMIndoorMap, and this works by
applying a custom OSM changeset file (.osc) on top of the base map data.
This allows e.g. adding transient event-specific content that cannot be
put in the OSM data directly, as well as changing event-specific properties
of OSM elements (renaming/repurposing, access limitations, etc).