Don't reset active conference when opening the selection page
This is trying to address some confusing behavior with Kongress.
Currently if you have an active conference, and click this it will
suddenly (and without warning) boot you into an invalid state where
some buttons are broken and do nothing - like the "Check for updates"
button.

Instead, let's avoid this whole problem by not resetting the active
conference when going to this page. Then you can stay on your active
conference by clicking something else in the sidebar, canceling the
action. This makes the UX feel a little bit better.