Rework favorite handling
Make favorite button text depends on whether the event is already a
favorite or not, by making EventController a normal Qml element instead
of a singleton.

Pass all parameters manually to ScheduleView and FavoritesView to avoid
getting uid and other important parameters of the event garbage
collected (or at least invalid).